adding additional goals to kumuluzee plugin
