+++ tools/loader/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>kumuluzee-tools</artifactId>
         <groupId>com.kumuluz.ee</groupId>
-        <version>2.3.0-SNAPSHOT</version>
+        <version>2.4.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
