+++ tools/maven-plugin/src/main/java/com/kumuluz/ee/maven/plugin/CopyDependenciesAndWebappMojo.java
@@ -1,0 +1,30 @@
+package com.kumuluz.ee.maven.plugin;
+
+import org.apache.maven.execution.MavenSession;
+import org.apache.maven.plugin.BuildPluginManager;
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugins.annotations.*;
+import org.apache.maven.project.MavenProject;
+
+@Mojo(
+        name = "copy-dependencies",
+        defaultPhase = LifecyclePhase.PACKAGE,
+        requiresDependencyResolution = ResolutionScope.COMPILE_PLUS_RUNTIME,
+        requiresDependencyCollection = ResolutionScope.COMPILE_PLUS_RUNTIME
+)
+public class CopyDependenciesAndWebappMojo extends AbstractCopyDependenciesAndWebappMojo {
+
+    @Parameter(defaultValue = "${project}", readonly = true, required = true)
+    private MavenProject mavenProject;
+
+    @Parameter(defaultValue = "${session}", readonly = true, required = true)
+    private MavenSession mavenSession;
+
+    @Component
+    private BuildPluginManager buildPluginManager;
+
+    @Override
+    public void execute() throws MojoExecutionException {
+        copyDependencies(/*mavenProject, mavenSession, buildPluginManager*/);
+    }
+}
